Fix calendar crash on device, fix charting frame
