+++ app/controllers/ios/collectionview.js
@@ -45,7 +45,7 @@
     
 	// Calculate the cell specs 
     var screenRect = UIScreen.mainScreen.bounds;
-    screenRect = CGRectMake(0 , 0, UIScreen.mainScreen.bounds.size.width, UIScreen.mainScreen.bounds.size.height - 64);
+    screenRect = CGRectMake(0, 0, UIScreen.mainScreen.bounds.size.width, UIScreen.mainScreen.bounds.size.height - 64);
     var cellWidth = screenRect.size.width / 3.0;
     
 	// Create the cell layout and assign the size
