+++ app/controllers/ios/libraries/charting.js
@@ -33,7 +33,7 @@
 
 	// make the chart take up most of the screen bounds
 	var bounds = UIScreen.mainScreen.bounds;
-	chart.frame = CGRectMake(0, 0, bounds.size.width - 40, bounds.size.height - 100);
+	chart.frame = CGRectMake(20, 0, bounds.size.width - 40, bounds.size.height - 100);
 
 	// docs say we need to reload the data initially to cause it to paint
 	chart.reloadData();
